move clearKubeconfigEnvVars to separate file
